DEV: Split controller up and add tests

+++ db/migrate/20260211000000_create_nested_view_post_stats.rb
@@ -3,7 +3,7 @@
 class CreateNestedViewPostStats < ActiveRecord::Migration[7.2]
   def change
     create_table :nested_view_post_stats do |t|
-      t.integer :post_id, null: false
+      t.bigint :post_id, null: false
       t.integer :direct_reply_count, default: 0, null: false
       t.timestamps
     end
